Add fonduer-specific snorkel base files
This is a copy of the snorkel directory (minus the contrib folder) from
the fonduer branch of https://github.com/hazyresearch/snorkel. We chose
to duplicate this code due to the changes that we made to this snorkel
base, rather than merge in new code to the snorkel base that is only
used by fonduer. This will also allow future fonduer-specific
modifications to this base.


Former-commit-id: a5263eb99cc68f11472ff7b2d4cd9488ee3958b7