Remove get parent/children/sentence generator from Context
Only a few of the classes inheriting from Context implemented these
functions, and they are not as clear as explicitly navigating the data
model. For example, the Span.get_parent() is less explicit than just
saying Span.sentence. Rather than prsent an interface that is not
explicit and rarely implemented, this commit just removes it.