Change import syntax to be more explicit
This refactor eliminates the host of imports that occurred at
fonduer/__init__.py and opts for more explicit import syntax to make it
clearer to the user which pipeline phases each import is associated
with. With this change, the public API of each pipeline phase is more
clearly defined in each pipeline's respective __init__.py.

In addition, this commit splits the old annotations.py file into two
separate files, one for the Annotator classes, and one of annotation
utilities.