fix(udf/parser): fix `get_last_documents` to so that it queries from the db
The previous implementation would track the last parsed documents using
a set of Document objects returned by a Preprocesser. Thus, when a user
called `get_last_documents`, they would get Documents, but not Documents
which had any linkage to the underlying database, and thus could not
view the document's ID, or relationships like Sentences.

This commit is twofold. First, we move the `get_last_documents` call
implementation to the Parser, rather than having it in the UDF. This
just makes more organizational sense. Second, it queries the database
for Documents by filtering on the names of the last documents, ensuring
that the returned documents can be navigated by the user (e.g., to view
the Sentences).