fix(cand_gen): fix cascading deletion of Candidates
We currently do not have relationships from mention_subclass to Mention,
or candidate_subclass to Candidate, meaning that deleting entries of the
subclasses does not delete the corresponding entries in the parent
class. This commit provides a workaround for the user API by clearing
the Candidate table when Mentions are cleared using the
MentionExtractor.

In an ideal world, we would have relationships set up that handle this
automatically.