Reorganize learning module (#203)
* rename layers to modules and add loss modules

* remove SoftCrossEntropyLoss in utils

* 1. remove original Classifier class, and rename NoiseAwareModel to Classifier; 2. use the same setting for both binary and multi-class classifier; 3. remove LabelBalancer; 4. code clean up

* 1. use fbeta instead of f1; 2. clean logging info

* update imports

* update score function

* 1. add roc aux metric; 2. make variable name clear

* 1. update e2e test; 2. fix issue

* reorg the learning code

* unify _calc_logits

* Update src/fonduer/learning/disc_models/lstm.py

Co-Authored-By: SenWu <SenWu@users.noreply.github.com>

* Update CHANGELOG.rst

Co-Authored-By: SenWu <SenWu@users.noreply.github.com>

* Update src/fonduer/learning/disc_models/sparse_logistic_regression.py

Co-Authored-By: SenWu <SenWu@users.noreply.github.com>

* Update src/fonduer/learning/disc_models/logistic_regression.py

Co-Authored-By: SenWu <SenWu@users.noreply.github.com>

* address comments

* Update src/fonduer/learning/disc_models/sparse_lstm.py

Co-Authored-By: SenWu <SenWu@users.noreply.github.com>

* docs(learning): remove mock Classifier to show inherited members

* cleanup format

* docs: specify Python 3.6 in ReadTheDocs

* cleanup format

* test(cand_gen): add test for cascading deletion of mentions to candidates

When a Mention is deleted, the mention_subclass, as the child, is
properly deleted. Similarly, the corresponding candidate_subclass is
properly deleted. However, the relationship from candidate_subclass to
the parent Candidate class does not cascade delete. That is, we have a
case where we see this behavior:

```
(Pdb) session.query(Candidate).count()
3878
(Pdb) session.query(PartTemp).count()
3877
(Pdb) session.query(Mention).filter(Mention.type == "part").delete()
234
(Pdb) session.query(Part).count()
0
(Pdb) session.query(PartTemp).count()
0
(Pdb) session.query(Candidate).count()
3878
```

The Candidate table should also be cleared after this procedure. This
commit as a test for this scenario. Importantly, the same is not true in
the other direction--deleting Candidates should NOT affect the Mentions.

* fix(cand_gen): fix cascading deletion of Candidates

We currently do not have relationships from mention_subclass to Mention,
or candidate_subclass to Candidate, meaning that deleting entries of the
subclasses does not delete the corresponding entries in the parent
class. This commit provides a workaround for the user API by clearing
the Candidate table when Mentions are cleared using the
MentionExtractor.

In an ideal world, we would have relationships set up that handle this
automatically.

* test(incremental): split incremental test into separate file w/ own DB

This isolates the incremental test from being affected by the side
effects from the e2e tests.

* docs(changelog): add fix for mention_extractor clear/clear_all to changelog