fix(visualizer): with with protections for wand objects
The wand documentation suggest to wrap the Color and Drawing objects in
a `with` block for more explicit resource management [1] [2].

To quote [1]:
> Unlike any other objects in Wand, its resource management can be
> implicit when it used outside of with block. In these case, its
> resource are allocated for every operation which requires a resource
> and destroyed immediately. Of course it is inefficient when the
> operations are much, so to avoid it, you should use color objects
> inside of with block explicitly

The example code for using Drawing() also uses the with block.

Note that we do NOT wrap the Image() object in a with block because that
must not be freed in visualizer.py.

[1]: http://docs.wand-py.org/en/0.5.9/wand/color.html
[2]: http://docs.wand-py.org/en/0.5.9/guide/draw.html