build(github-actions): fix tests on macOS with OpenMP conflicts
In recent tests on both Travis-CI and Github Actions, we found that
running test_e2e.py and test_visualizer.py together (regardless of other
tests that may also have been run) caused an `Abort Trap 6` error and
crashed the tests, causing a failure. This only occured on macOS, not
linux.

By ssh-ing into the test VM, and using log-cli with DEBUG-level logging,
we found this error message:

    OMP: Error #15: Initializing libomp.dylib, but found libiomp5.dylib
    already initialized.
    OMP: Hint This means that multiple copies of the OpenMP runtime have
    been linked into the program. That is dangerous, since it can
    degrade performance or cause incorrect results. The best thing to do
    is to ensure that only a single OpenMP runtime is linked into the
    process, e.g. by avoiding static linking of the OpenMP runtime in
    any library. As an unsafe, unsupported, undocumented workaround you
    can set the environment variable KMP_DUPLICATE_LIB_OK=TRUE to allow
    the program to continue to execute, but that may cause crashes or
    silently produce incorrect results. For more information, please see
    http://openmp.llvm.org/
    Fatal Python error: Aborted

There are a few workarounds suggested in various forums [1] [2], but as
we aren't using Conda, we opt for the workaround of just setting the
environment variable, which appears to solve the problem.

[1]: https://stackoverflow.com/questions/53014306/error-15-initializing-libiomp5-dylib-but-found-libiomp5-dylib-already-initial
[2]: https://github.com/dmlc/xgboost/issues/1715