build(github-actions): fix macOS tests by using default python (#386)
On macOS, Python installed by default (3.7) is used.
actions/setup-python@v1 cannot be used on macos because "Can't connect to HTTPS URL because the SSL module is not available." error happens when trying to pip install.
Not confirmed, but this is probably because brew install openssl@1.1 as imagemagick dependency and breaks a link Python installed actions/setup-python@v1 and openssl.

For simplicity, and because we test Python 3.6 on linux, this drops Python 3.6 tests on macOS.