feat(packaging): add a custom MLflow model for Fonduer (#407)
Users would like to be able to develop a Fonduer-based application
locally on their workstation, and then package the whole pipeline
(parsing, extraction, featurization, and classification) to be deployed
somewhere remote to serve. However, FOnduer-based applications are not
easily packaged.

This commit changes that. Specifically, this creates a serializable
Fonduer model that is capable of executing any phase of the Fonduer
pipeline. It is manageable by MLflow [1], a platform for managing machine
learning pipelines, using the MLflow Model as a storage format.

Further documentation has been added in docs/user/packaging.rst, and
usage is also shown in the additional tests.

[1]: https://mlflow.org/

Closes #259.